Add I/O progress helper, and stop using the goftp library.
The version of vsftpd used by CloudControl has a slightly odd  interpretation of FTPS; it REQUIRES that the same TLS session be reused across the control and data connections.

Unfortunately I don't know of any FTP client libraries for Go that actually work this way, so we'll have to stick to using curl for now.